all new subcomponents working now
